KNN/requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
colorama==0.4.6
filelock==3.20.0
fsspec==2025.12.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.6.1
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
pillow==12.0.0
pydantic==2.12.5
pydantic_core==2.41.5
PyYAML==6.0.3
scikit-learn==1.8.0
scipy==1.16.3
setuptools==70.2.0
simsimd==6.5.8
stringzilla==4.5.1
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.9.1+cpu
torchvision==0.24.1+cpu
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
